requirements.txt
twitchAPI
sanic
python-dotenv